EES-3372 - slight cleanup of shared refres =h token code
